PM2020T9-35

+++ mihraz7/tendersOffers/views.py
@@ -9,12 +9,16 @@ def RegisterOffer(request):
     lastName = request.user.last_name
     Email = request.user.email
     Offer = request.POST.get("Offer")
-    
+    count = tenderModel.Tender.objects.get(tender_id=request.POST.get("tenId")).Count_of_applied
+    tenderModel.Tender.objects.filter(tender_id=tenderId).update(Count_of_applied=count+1)
     newOffer = models.TenderOffer(tender_id=tenderId,tender_name=tenderName,first_name=firstName,last_name=lastName,email=Email,offer=Offer)
     newOffer.save()
 
 def DeleteOffer(request):
-    models.TenderOffer.objects.filter(tender_id=request.POST.get("tenIdDelete"),email=request.user.email).delete()
+    if(models.TenderOffer.objects.filter(tender_id=request.POST.get("tenIdDelete"),email=request.user.email).count()>0):
+        count = tenderModel.Tender.objects.get(tender_id=request.POST.get("tenIdDelete")).Count_of_applied
+        tenderModel.Tender.objects.filter(tender_id=request.POST.get("tenIdDelete")).update(Count_of_applied=count-models.TenderOffer.objects.filter(tender_id=request.POST.get("tenIdDelete"),email=request.user.email).count())
+        models.TenderOffer.objects.filter(tender_id=request.POST.get("tenIdDelete"),email=request.user.email).delete()
 
 def AllTendersOffers():
     return models.TenderOffer.objects.all()
